backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.0.0
Pillow==10.1.0

